Auto generate statement nodes
